pyproject.toml
[project]
name = "travel-log"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "atsushi", email = "green515tea@gmail.com" }
]
dependencies = [
    "piexif>=1.1.3",
    "Pillow>=10.4.0",
    "pillow_heif>=0.18.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/travel_log"]
